renaming and api name changes
